ongoing work on re-shaping and refactor py-wrapping and using ctest
