+++ CMakePresets.json
@@ -6,7 +6,7 @@
       "generator": "Ninja",
       "hidden": true,
       "cacheVariables": {
-        "CMAKE_PREFIX_PATH": "$env{LIBRARY_PREFIX};$env{LIBRARY_PREFIX}/include;$env{LIBRARY_PREFIX}/lib;$env{LIBRARY_PREFIX}/bin",
+        "CMAKE_PREFIX_PATH": "$env{LIBRARY_PREFIX};$env{LIBRARY_PREFIX}/include;$env{PREFIX}/include;$env{LIBRARY_PREFIX}/lib;$env{LIBRARY_PREFIX}/bin",
         "CMAKE_INSTALL_PREFIX": "$env{LIBRARY_PREFIX}",
         "CGAL_DIR": "$env{LIBRARY_PREFIX}/lib/cmake/CGAL",
         "OpenCASCADE_DIR": "$env{LIBRARY_PREFIX}/lib/cmake/opencascade",
@@ -44,8 +44,7 @@
         "CMAKE_BUILD_TYPE": "RelwithDebInfo",
         "BUILD_TESTING": "ON",
         "BUILD_PYTHON": "ON",
-        "BUILD_STP2GLB": "ON",
-        "CONDA_LOCAL_DEV": "ON"
+        "BUILD_STP2GLB": "ON"
       }
     },
     {
@@ -95,7 +94,6 @@
         "OpenCASCADE_INCLUDE_DIR": "$env{LIBRARY_PREFIX}/include/opencascade",
         "Python_EXECUTABLE": "$env{PREFIX}/bin/python",
         "BUILD_TESTING": "ON",
-        "CONDA_LOCAL_DEV": "OFF",
         "BUILD_WASM": "ON"
       }
     },
@@ -114,7 +112,6 @@
         "OpenCASCADE_DIR": "$env{LIBRARY_PREFIX}/lib/cmake/opencascade",
         "OpenCASCADE_INCLUDE_DIR": "$env{LIBRARY_PREFIX}/include/opencascade",
         "BUILD_TESTING": "ON",
-        "CONDA_LOCAL_DEV": "OFF",
         "BUILD_WASM": "OFF"
       }
     }
