+++ cmake/build_stp2glb.cmake
@@ -50,5 +50,5 @@ if (BUILD_TESTING)
     message(STATUS "Building the testing tree.")
     set(CMAKE_INSTALL_RPATH "${STP2GLB_BIN_DIR}")
     set(CMAKE_BUILD_RPATH "${STP2GLB_BIN_DIR}")
-    include(tests/cpp/tests.cmake)
+    include(tests/cpp/stp2glb_tests.cmake)
 endif (BUILD_TESTING)
