+++ src/cadit/tinygltf/tiny.h
@@ -1,10 +1,6 @@
-//
-// Created by Kristoffer on 26/07/2023.
-//
-
-
 #ifndef ADA_CPP_TINY_H
 #define ADA_CPP_TINY_H
+
 #include "../../binding_core.h"
 
 void tiny_gltf_module(nb::module_&m);
