+++ src/cadit/tinygltf/tinygltf.cpp
@@ -1,10 +1,4 @@
-//
-// Created by Kristoffer on 23/01/2024.
-//
-#pragma once
-
 #define TINYGLTF_IMPLEMENTATION
 #define STB_IMAGE_IMPLEMENTATION
 #define STB_IMAGE_WRITE_IMPLEMENTATION
-
 #include "tiny_gltf.h"
