+++ src/helpers/helpers.h
@@ -2,14 +2,10 @@
 #define NANO_OCCT_HELPERS_H
 
 #include <vector>
-#include <gp_XYZ.hxx>
-#include <gp_Pnt.hxx>
 #include <TopoDS_Solid.hxx>
 #include <BRepPrimAPI_MakeBox.hxx>
 #include <TDF_Label.hxx>
-#include <TDataStd_Name.hxx>
 #include <Quantity_Color.hxx>
-#include <XCAFDoc_ColorType.hxx>
 #include <XCAFDoc_ColorTool.hxx>
 #include <optional>
 #include "../geom/Color.h"
