+++ src/visit/tess_py_wrap.h
@@ -1,0 +1,6 @@
+//
+// Created by ofskrand on 24.01.2025.
+//
+
+#include "../binding_core.h"
+void tess_helper_module(nb::module_ &m);
